refactored redis and mongoose connections, also split out app from server
