+++ config/default.json
@@ -3,6 +3,6 @@
       "url": "mongodb://localhost:27017/efuseMongo"
     },
     "redis": {
-      "url": "redis://localhost:6379"
+      "url": "redis://127.0.0.1:6379"
     }
 }
